graph: move storage things to storage.go
